move DrawingSub, StoryViewer, WritingSub components to MissionControl directory, remove unneded container componentts, indexes, directories
