--- src/components/pages/DrawingSub/index.js
@@ -1,1 +1,0 @@
-export { default as DrawingSub } from './DrawingSubContainer';
