--- src/components/pages/StoryPrompt/index.js
@@ -1,1 +1,0 @@
-export { default as StoryPrompt } from './StoryViewerContainer';
