--- src/components/pages/WritingSub/WritingSubContainer.js
@@ -1,47 +1,0 @@
-import React, { useState, useEffect, useMemo } from 'react';
-import { useOktaAuth } from '@okta/okta-react';
-
-import RenderWritingSub from './RenderWritingSub';
-
-function WritingSubContainer({ LoadingComponent, ...props }) {
-  const { authState, authService } = useOktaAuth();
-  const [userInfo, setUserInfo] = useState(null);
-  // eslint-disable-next-line
-  const [memoAuthService] = useMemo(() => [authService], []);
-
-  useEffect(() => {
-    let isSubscribed = true;
-
-    memoAuthService
-      .getUser()
-      .then(info => {
-        // if user is authenticated we can use the authService to snag some user info.
-        // isSubscribed is a boolean toggle that we're using to clean up our useEffect.
-        if (isSubscribed) {
-          setUserInfo(info);
-        }
-      })
-      .catch(err => {
-        isSubscribed = false;
-        return setUserInfo(null);
-      });
-    return () => (isSubscribed = false);
-  }, [memoAuthService]);
-
-  return (
-    <>
-      {authState.isAuthenticated && !userInfo && (
-        <LoadingComponent message="Loading..." />
-      )}
-      {authState.isAuthenticated && userInfo && (
-        <RenderWritingSub
-          {...props}
-          userInfo={userInfo}
-          authService={authService}
-        />
-      )}
-    </>
-  );
-}
-
-export default WritingSubContainer;
