--- src/components/pages/WritingSub/index.js
@@ -1,1 +1,0 @@
-export { default as WritingSub } from './WritingSubContainer';
